feat: Chats sync with different users

+++ convex/chat.ts
@@ -1,11 +1,25 @@
+import { Id } from "@/convex/_generated/dataModel";
 import { mutation, query } from "@/convex/_generated/server";
+import { getAuthUserId } from "@convex-dev/auth/server";
 import { v } from "convex/values";
 
 export const createChat = mutation({
   args: { title: v.string() },
   handler: async (ctx, { title }) => {
+    const userId = await getAuthUserId(ctx);
+
+    if (!userId) {
+      throw new Error("User not authenticated");
+    }
+
+    const user = await ctx.db.get(userId as Id<"users">);
+    if (!user) {
+      throw new Error("User not found");
+    }
+
     const chatId = await ctx.db.insert("chat", {
       title,
+      userId: user._id,
     });
 
     return chatId;
@@ -14,17 +28,41 @@ export const createChat = mutation({
 
 export const getChats = query({
   handler: async (ctx) => {
-    const chats = await ctx.db.query("chat").collect();
-    return chats;
+    const userId = await getAuthUserId(ctx);
+    if (!userId) {
+      throw new Error("User not authenticated");
+    }
+
+    const user = await ctx.db.get(userId as Id<"users">);
+    if (!user) {
+      throw new Error("User not found");
+    }
+
+    const chats = await ctx.db
+      .query("chat")
+      .withIndex("userId", (q) => q.eq("userId", user._id))
+      .collect();
+    return chats.reverse();
   },
 });
 
 export const renameChat = mutation({
   args: { chatId: v.id("chat"), title: v.string() },
   handler: async (ctx, { chatId, title }) => {
+    const userId = await getAuthUserId(ctx);
+    if (!userId) {
+      throw new Error("User not authenticated");
+    }
+    const user = await ctx.db.get(userId as Id<"users">);
+    if (!user) {
+      throw new Error("User not found");
+    }
     const chat = await ctx.db.get(chatId);
     if (!chat) {
       throw new Error("Chat not found");
+    }
+    if (chat.userId !== user._id) {
+      throw new Error("User does not have permission to rename this chat");
     }
     await ctx.db.patch(chatId, { title });
     return chatId;
@@ -34,15 +72,21 @@ export const renameChat = mutation({
 export const deleteChatAndMessages = mutation({
   args: { chatId: v.id("chat") },
   handler: async (ctx, { chatId }) => {
-    const messages = await ctx.db
-      .query("message")
-      .filter((q) => q.eq(q.field("chatId"), chatId))
-      .collect();
-
-    for (const message of messages) {
-      await ctx.db.delete(message._id);
+    const userId = await getAuthUserId(ctx);
+    if (!userId) {
+      throw new Error("User not authenticated");
     }
-
+    const user = await ctx.db.get(userId as Id<"users">);
+    if (!user) {
+      throw new Error("User not found");
+    }
+    const chat = await ctx.db.get(chatId);
+    if (!chat) {
+      throw new Error("Chat not found");
+    }
+    if (chat.userId !== user._id) {
+      throw new Error("User does not have permission to delete this chat");
+    }
     await ctx.db.delete(chatId);
 
     return true;
